Separated simulation class from main program
